prepare-database/requirements.txt
astroid==2.5.6
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
idna==2.10
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysql-connector==2.2.9
numpy==1.19.5
pandas==1.1.5
pyarrow==4.0.1
pycodestyle==2.7.0
pylint==2.8.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
urllib3==1.26.5
wrapt==1.12.1
